fix: missing icons (#2269)
